+++ crates/zng/src/toggle.rs
@@ -181,8 +181,8 @@
 
 pub use zng_wgt_toggle::{
     check_spacing, combo_spacing, deselect_on_deinit, deselect_on_new, is_checked, radio_spacing, scroll_on_select, select_on_init,
-    select_on_new, selector, style_fn, switch_spacing, tristate, CheckStyle, ComboStyle, DefaultStyle, RadioStyle, Selector, SelectorError,
-    SelectorImpl, SwitchStyle, Toggle, IS_CHECKED_VAR,
+    select_on_new, selector, style_fn, switch_spacing, tristate, CheckStyle, ComboStyle, DefaultStyle, LightStyle, RadioStyle, Selector,
+    SelectorError, SelectorImpl, SwitchStyle, Toggle, IS_CHECKED_VAR,
 };
 
 /// Toggle commands.