+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `button::LightStyle!()` and `toggle::LightStyle!()`.
 * Fix when expr not recognized.
 * Fix `WINDOWS.is_loading`.
 * Add `WINDOWS.wait_loaded`.